+++ mgba-android-memapi/android-fork/app/src/main/java/hh/game/mgba_android/tracker/quickload/RomFamily.kt
@@ -1,5 +1,7 @@
 package hh.game.mgba_android.tracker.quickload
 
+enum class FamilyMode { BATCH, VANILLA }
+
 /**
  * Parsed representation of a ROM filename.
  *